refactor project details
